@@ -1,14 +1,13 @@
 ﻿using System;
 
-namespace BattleChess3.Core.Model.Figures
+namespace BattleChess3.Core.Model.Figures;
+
+[Flags]
+public enum FigureTypes
 {
-    [Flags]
-    public enum FigureTypes
-    {
-        Nothing = 0,
-        Foot = 1,
-        Mount = 2,
-        Special = 4,
-        Object = 8
-    }
+    Nothing = 0,
+    Foot = 1,
+    Mount = 2,
+    Special = 4,
+    Object = 8
 }
