@@ -1,15 +1,14 @@
-﻿namespace BattleChess3.Core.Model.Figures
+﻿namespace BattleChess3.Core.Model.Figures;
+
+public interface IFigureGroup
 {
-    public interface IFigureGroup
-    {
-        /// <summary>
-        /// Shown name of figure group
-        /// </summary>
-        string ShownName { get; }
-        
-        /// <summary>
-        /// Figure types of group
-        /// </summary>
-        IFigureType[] FigureTypes { get; }
-    }
+    /// <summary>
+    /// Shown name of figure group
+    /// </summary>
+    string ShownName { get; }
+    
+    /// <summary>
+    /// Figure types of group
+    /// </summary>
+    IFigureType[] FigureTypes { get; }
 }
