@@ -1,13 +1,12 @@
-﻿namespace BattleChess3.UI
+﻿namespace BattleChess3.UI;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App
 {
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App
+    public App()
     {
-        public App()
-        {
-            DependenciesBuilder.Initialize();
-        }
+        DependenciesBuilder.Initialize();
     }
 }
