@@ -3,22 +3,21 @@ using System.Windows.Controls;
 using System.Windows.Documents;
 using System.Windows.Media;
 
-namespace BattleChess3.UI.Views
-{
-    public partial class MainWindow
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-            TextElement.FontFamilyProperty.OverrideMetadata(
-                typeof(TextElement),
-                new FrameworkPropertyMetadata(
-                    new FontFamily("Britannic Bold")));
+namespace BattleChess3.UI.Views;
 
-            TextBlock.FontFamilyProperty.OverrideMetadata(
-                typeof(TextBlock),
-                new FrameworkPropertyMetadata(
-                    new FontFamily("Britannic Bold")));
-        }
+public partial class MainWindow
+{
+    public MainWindow()
+    {
+        InitializeComponent();
+        TextElement.FontFamilyProperty.OverrideMetadata(
+            typeof(TextElement),
+            new FrameworkPropertyMetadata(
+                new FontFamily("Britannic Bold")));
+
+        TextBlock.FontFamilyProperty.OverrideMetadata(
+            typeof(TextBlock),
+            new FrameworkPropertyMetadata(
+                new FontFamily("Britannic Bold")));
     }
 }
