@@ -5,60 +5,59 @@ using BattleChess3.Core.Model;
 using BattleChess3.Core.Model.Figures;
 using BattleChess3.DefaultFigures.Utilities;
 
-namespace BattleChess3.ChessFigures
+namespace BattleChess3.ChessFigures;
+
+public class King : IFigureType
 {
-    public class King : IFigureType
+    public static readonly King Instance = new();
+    public string ShownName => CurrentLocalization.Instance["King_Name"];
+    public string Description => CurrentLocalization.Instance["King_Description"];
+    public string UnitName => $"{nameof(ChessFigureGroup)}.{nameof(King)}";
+    public FigureTypes UnitTypes => FigureTypes.Special;
+    public FigureTypes Bonus => FigureTypes.Nothing;
+    public FigureTypes AntiBonus => FigureTypes.Nothing;
+    public double FullHp => 100;
+    public double Attack => 100;
+    public double Defence => 0;
+    public bool MovingAttack => true;
+    public int Cost => 0;
+
+    public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
     {
-        public static readonly King Instance = new King();
-        public string ShownName => CurrentLocalization.Instance["King_Name"];
-        public string Description => CurrentLocalization.Instance["King_Description"];
-        public string UnitName { get; } = $"{nameof(ChessFigureGroup)}.{nameof(King)}";
-        public FigureTypes UnitTypes { get; } = FigureTypes.Special;
-        public FigureTypes Bonus { get; } = FigureTypes.Nothing;
-        public FigureTypes AntiBonus { get; } = FigureTypes.Nothing;
-        public double FullHp { get; } = 100;
-        public double Attack { get; } = 100;
-        public double Defence { get; } = 0;
-        public bool MovingAttack { get; } = true;
-        public int Cost { get; } = 0;
+        {0, new Uri("pack://application:,,,/BattleChess3.ChessFigures;component/Images/King1.png", UriKind.Absolute)},
+        {1, new Uri("pack://application:,,,/BattleChess3.ChessFigures;component/Images/King2.png", UriKind.Absolute)},
+    };
 
-        public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
-        {
-            {0, new Uri("pack://application:,,,/BattleChess3.ChessFigures;component/Images/King1.png", UriKind.Absolute)},
-            {1, new Uri("pack://application:,,,/BattleChess3.ChessFigures;component/Images/King2.png", UriKind.Absolute)},
-        };
+    public void AttackAction(ITile from, ITile to, ITile[] board)
+        => to.KillFigure(board);
 
-        public void AttackAction(ITile from, ITile to, ITile[] board)
-            => to.KillFigure(board);
+    public void MoveAction(ITile from, ITile to, ITile[] board)
+        => from.MoveToPosition(to.Position, board);
 
-        public void MoveAction(ITile from, ITile to, ITile[] board)
-            => from.MoveToPosition(to.Position, board);
-
-        private readonly Position[][] _moveChain = 
-        {
-            new Position[] {(1, 1)},
-            new Position[] {(1, 0)},
-            new Position[] {(1, -1)},
-            new Position[] {(0, 1)},
-            new Position[] {(0, -1)},
-            new Position[] {(-1, 1)},
-            new Position[] {(-1, 0)},
-            new Position[] {(-1, -1)},
-        };
-        public Position[][] GetMoveChains(Position position) => _moveChain;
-        
-        
-        private readonly Position[][] _attackChain = 
-        {
-            new Position[] {(1, 1)},
-            new Position[] {(1, 0)},
-            new Position[] {(1, -1)},
-            new Position[] {(0, 1)},
-            new Position[] {(0, -1)},
-            new Position[] {(-1, 1)},
-            new Position[] {(-1, 0)},
-            new Position[] {(-1, -1)},
-        };
-        public Position[][] GetAttackChains(Position position) => _attackChain;
-    }
+    private readonly Position[][] _moveChain = 
+    {
+        new Position[] {(1, 1)},
+        new Position[] {(1, 0)},
+        new Position[] {(1, -1)},
+        new Position[] {(0, 1)},
+        new Position[] {(0, -1)},
+        new Position[] {(-1, 1)},
+        new Position[] {(-1, 0)},
+        new Position[] {(-1, -1)},
+    };
+    public Position[][] GetMoveChains(Position position) => _moveChain;
+    
+    
+    private readonly Position[][] _attackChain = 
+    {
+        new Position[] {(1, 1)},
+        new Position[] {(1, 0)},
+        new Position[] {(1, -1)},
+        new Position[] {(0, 1)},
+        new Position[] {(0, -1)},
+        new Position[] {(-1, 1)},
+        new Position[] {(-1, 0)},
+        new Position[] {(-1, -1)},
+    };
+    public Position[][] GetAttackChains(Position position) => _attackChain;
 }
