@@ -5,52 +5,51 @@ using BattleChess3.Core.Model.Figures;
 using BattleChess3.DefaultFigures.Utilities;
 using BattleChess3.HobbitFigures.Localization;
 
-namespace BattleChess3.HobbitFigures
+namespace BattleChess3.HobbitFigures;
+
+public class MinorWizard : IFigureType
 {
-    public class MinorWizard : IFigureType
+    public static readonly MinorWizard Instance = new();
+    public string ShownName => CurrentLocalization.Instance["MinorWizard_Name"];
+    public string Description => CurrentLocalization.Instance["MinorWizard_Description"];
+    public string UnitName => $"{nameof(HobbitFigureGroup)}.{nameof(MinorWizard)}";
+    public FigureTypes UnitTypes => FigureTypes.Foot;
+    public FigureTypes Bonus => FigureTypes.Nothing;
+    public FigureTypes AntiBonus => FigureTypes.Nothing;
+    public double FullHp => 100;
+    public double Attack => 100;
+    public double Defence => 0;
+    public bool MovingAttack => true;
+    public int Cost => 5;
+
+    public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
     {
-        public static readonly MinorWizard Instance = new MinorWizard();
-        public string ShownName => CurrentLocalization.Instance["MinorWizard_Name"];
-        public string Description => CurrentLocalization.Instance["MinorWizard_Description"];
-        public string UnitName { get; } = $"{nameof(HobbitFigureGroup)}.{nameof(MinorWizard)}";
-        public FigureTypes UnitTypes { get; } = FigureTypes.Foot;
-        public FigureTypes Bonus { get; } = FigureTypes.Nothing;
-        public FigureTypes AntiBonus { get; } = FigureTypes.Nothing;
-        public double FullHp { get; } = 100;
-        public double Attack { get; } = 100;
-        public double Defence { get; } = 0;
-        public bool MovingAttack { get; } = true;
-        public int Cost { get; } = 5;
+        {0, new Uri("pack://application:,,,/BattleChess3.HobbitFigures;component/Images/MinorWizard1.png", UriKind.Absolute)},
+        {1, new Uri("pack://application:,,,/BattleChess3.HobbitFigures;component/Images/MinorWizard2.png", UriKind.Absolute)},
+    };
 
-        public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
-        {
-            {0, new Uri("pack://application:,,,/BattleChess3.HobbitFigures;component/Images/MinorWizard1.png", UriKind.Absolute)},
-            {1, new Uri("pack://application:,,,/BattleChess3.HobbitFigures;component/Images/MinorWizard2.png", UriKind.Absolute)},
-        };
+    public void AttackAction(ITile from, ITile to, ITile[] board)
+        => to.KillFigure(board);
 
-        public void AttackAction(ITile from, ITile to, ITile[] board)
-            => to.KillFigure(board);
+    public void MoveAction(ITile from, ITile to, ITile[] board)
+        => from.MoveToPosition(to.Position, board);
 
-        public void MoveAction(ITile from, ITile to, ITile[] board)
-            => from.MoveToPosition(to.Position, board);
-
-        private readonly Position[][] _moveChain = 
-        {
-            new Position[] {(1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0)}, 
-            new Position[] {(0, 1), (0, 2), (0, 3), (0, 4), (0, 5), (0, 6), (0, 7)},
-            new Position[] {(-1, 0), (-2, 0), (-3, 0), (-4, 0), (-5, 0), (-6, 0), (-7, 0)},
-            new Position[] {(0, -1), (0, -2), (0, -3), (0, -4), (0, -5), (0, -6), (0, -7)}
-        };
-        public Position[][] GetMoveChains(Position position) => _moveChain;
-        
-        
-        private readonly Position[][] _attackChain = 
-        {
-            new Position[] {(1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0)}, 
-            new Position[] {(0, 1), (0, 2), (0, 3), (0, 4), (0, 5), (0, 6), (0, 7)},
-            new Position[] {(-1, 0), (-2, 0), (-3, 0), (-4, 0), (-5, 0), (-6, 0), (-7, 0)},
-            new Position[] {(0, -1), (0, -2), (0, -3), (0, -4), (0, -5), (0, -6), (0, -7)}
-        };
-        public Position[][] GetAttackChains(Position position) => _attackChain;
-    }
+    private readonly Position[][] _moveChain = 
+    {
+        new Position[] {(1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0)}, 
+        new Position[] {(0, 1), (0, 2), (0, 3), (0, 4), (0, 5), (0, 6), (0, 7)},
+        new Position[] {(-1, 0), (-2, 0), (-3, 0), (-4, 0), (-5, 0), (-6, 0), (-7, 0)},
+        new Position[] {(0, -1), (0, -2), (0, -3), (0, -4), (0, -5), (0, -6), (0, -7)}
+    };
+    public Position[][] GetMoveChains(Position position) => _moveChain;
+    
+    
+    private readonly Position[][] _attackChain = 
+    {
+        new Position[] {(1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0)}, 
+        new Position[] {(0, 1), (0, 2), (0, 3), (0, 4), (0, 5), (0, 6), (0, 7)},
+        new Position[] {(-1, 0), (-2, 0), (-3, 0), (-4, 0), (-5, 0), (-6, 0), (-7, 0)},
+        new Position[] {(0, -1), (0, -2), (0, -3), (0, -4), (0, -5), (0, -6), (0, -7)}
+    };
+    public Position[][] GetAttackChains(Position position) => _attackChain;
 }
