@@ -5,60 +5,59 @@ using BattleChess3.Core.Model.Figures;
 using BattleChess3.DefaultFigures.Utilities;
 using BattleChess3.LordOfTheRingsFigures.Localization;
 
-namespace BattleChess3.LordOfTheRingsFigures
+namespace BattleChess3.LordOfTheRingsFigures;
+
+public class AragornSauron : IFigureType
 {
-    public class AragornSauron : IFigureType
+    public static readonly AragornSauron Instance = new();
+    public string ShownName => CurrentLocalization.Instance["AragornSauron_Name"];
+    public string Description => CurrentLocalization.Instance["AragornSauron_Description"];
+    public string UnitName => $"{nameof(LordOfTheRingsFigureGroup)}.{nameof(AragornSauron)}";
+    public FigureTypes UnitTypes => FigureTypes.Foot;
+    public FigureTypes Bonus => FigureTypes.Nothing;
+    public FigureTypes AntiBonus => FigureTypes.Nothing;
+    public double FullHp => 100;
+    public double Attack => 100;
+    public double Defence => 0;
+    public bool MovingAttack => true;
+    public int Cost => 0;
+
+    public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
     {
-        public static readonly AragornSauron Instance = new AragornSauron();
-        public string ShownName => CurrentLocalization.Instance["AragornSauron_Name"];
-        public string Description => CurrentLocalization.Instance["AragornSauron_Description"];
-        public string UnitName { get; } = $"{nameof(LordOfTheRingsFigureGroup)}.{nameof(AragornSauron)}";
-        public FigureTypes UnitTypes { get; } = FigureTypes.Foot;
-        public FigureTypes Bonus { get; } = FigureTypes.Nothing;
-        public FigureTypes AntiBonus { get; } = FigureTypes.Nothing;
-        public double FullHp { get; } = 100;
-        public double Attack { get; } = 100;
-        public double Defence { get; } = 0;
-        public bool MovingAttack { get; } = true;
-        public int Cost { get; } = 0;
+        {0, new Uri("pack://application:,,,/BattleChess3.LordOfTheRingsFigures;component/Images/AragornSauron1.png", UriKind.Absolute)},
+        {1, new Uri("pack://application:,,,/BattleChess3.LordOfTheRingsFigures;component/Images/AragornSauron2.png", UriKind.Absolute)},
+    };
 
-        public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
-        {
-            {0, new Uri("pack://application:,,,/BattleChess3.LordOfTheRingsFigures;component/Images/AragornSauron1.png", UriKind.Absolute)},
-            {1, new Uri("pack://application:,,,/BattleChess3.LordOfTheRingsFigures;component/Images/AragornSauron2.png", UriKind.Absolute)},
-        };
+    public void AttackAction(ITile from, ITile to, ITile[] board)
+        => to.KillFigure(board);
 
-        public void AttackAction(ITile from, ITile to, ITile[] board)
-            => to.KillFigure(board);
+    public void MoveAction(ITile from, ITile to, ITile[] board)
+        => from.MoveToPosition(to.Position, board);
 
-        public void MoveAction(ITile from, ITile to, ITile[] board)
-            => from.MoveToPosition(to.Position, board);
-
-        private readonly Position[][] _moveChain = 
-        {
-            new Position[] {(1, 1)},
-            new Position[] {(1, 0)},
-            new Position[] {(1, -1)},
-            new Position[] {(0, 1)},
-            new Position[] {(0, -1)},
-            new Position[] {(-1, 1)},
-            new Position[] {(-1, 0)},
-            new Position[] {(-1, -1)},
-        };
-        public Position[][] GetMoveChains(Position position) => _moveChain;
-        
-        
-        private readonly Position[][] _attackChain = 
-        {
-            new Position[] {(1, 1)},
-            new Position[] {(1, 0)},
-            new Position[] {(1, -1)},
-            new Position[] {(0, 1)},
-            new Position[] {(0, -1)},
-            new Position[] {(-1, 1)},
-            new Position[] {(-1, 0)},
-            new Position[] {(-1, -1)},
-        };
-        public Position[][] GetAttackChains(Position position) => _attackChain;
-    }
+    private readonly Position[][] _moveChain = 
+    {
+        new Position[] {(1, 1)},
+        new Position[] {(1, 0)},
+        new Position[] {(1, -1)},
+        new Position[] {(0, 1)},
+        new Position[] {(0, -1)},
+        new Position[] {(-1, 1)},
+        new Position[] {(-1, 0)},
+        new Position[] {(-1, -1)},
+    };
+    public Position[][] GetMoveChains(Position position) => _moveChain;
+    
+    
+    private readonly Position[][] _attackChain = 
+    {
+        new Position[] {(1, 1)},
+        new Position[] {(1, 0)},
+        new Position[] {(1, -1)},
+        new Position[] {(0, 1)},
+        new Position[] {(0, -1)},
+        new Position[] {(-1, 1)},
+        new Position[] {(-1, 0)},
+        new Position[] {(-1, -1)},
+    };
+    public Position[][] GetAttackChains(Position position) => _attackChain;
 }
