@@ -1,23 +1,22 @@
 ﻿using BattleChess3.Core.Model.Figures;
 using BattleChess3.LordOfTheRingsFigures.Localization;
 
-namespace BattleChess3.LordOfTheRingsFigures
-{
-    public class LordOfTheRingsFigureGroup : IFigureGroup
-    {
-        public string ShownName => CurrentLocalization.Instance["LordOfTheRingsFigureGroup_Name"];
+namespace BattleChess3.LordOfTheRingsFigures;
 
-        public IFigureType[] FigureTypes => new IFigureType[]
-        {
-            AragornSauron.Instance,
-            GandalfWitchKing.Instance,
-            FrodoGollum.Instance,
-            GimliNazgul.Instance,
-            LegolasNazgul.Instance,
-            MerryTroll.Instance,
-            PipinTroll.Instance,
-            SoldierOrc.Instance,
-            SamSaruman.Instance,
-        };
-    }
+public class LordOfTheRingsFigureGroup : IFigureGroup
+{
+    public string ShownName => CurrentLocalization.Instance["LordOfTheRingsFigureGroup_Name"];
+
+    public IFigureType[] FigureTypes => new IFigureType[]
+    {
+        AragornSauron.Instance,
+        GandalfWitchKing.Instance,
+        FrodoGollum.Instance,
+        GimliNazgul.Instance,
+        LegolasNazgul.Instance,
+        MerryTroll.Instance,
+        PipinTroll.Instance,
+        SoldierOrc.Instance,
+        SamSaruman.Instance,
+    };
 }
