@@ -5,60 +5,59 @@ using BattleChess3.Core.Model.Figures;
 using BattleChess3.DefaultFigures.Utilities;
 using BattleChess3.SilmarillionFigures.Localization;
 
-namespace BattleChess3.SilmarillionFigures
+namespace BattleChess3.SilmarillionFigures;
+
+public class IrmoUngoliant : IFigureType
 {
-    public class IrmoUngoliant : IFigureType
+    public static readonly IrmoUngoliant Instance = new();
+    public string ShownName => CurrentLocalization.Instance["IrmoUngoliant_Name"];
+    public string Description => CurrentLocalization.Instance["IrmoUngoliant_Description"];
+    public string UnitName => $"{nameof(SilmarillionFigureGroup)}.{nameof(IrmoUngoliant)}";
+    public FigureTypes UnitTypes => FigureTypes.Foot;
+    public FigureTypes Bonus => FigureTypes.Nothing;
+    public FigureTypes AntiBonus => FigureTypes.Nothing;
+    public double FullHp => 100;
+    public double Attack => 100;
+    public double Defence => 0;
+    public bool MovingAttack => true;
+    public int Cost => 3;
+
+    public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
     {
-        public static readonly IrmoUngoliant Instance = new IrmoUngoliant();
-        public string ShownName => CurrentLocalization.Instance["IrmoUngoliant_Name"];
-        public string Description => CurrentLocalization.Instance["IrmoUngoliant_Description"];
-        public string UnitName { get; } = $"{nameof(SilmarillionFigureGroup)}.{nameof(IrmoUngoliant)}";
-        public FigureTypes UnitTypes { get; } = FigureTypes.Foot;
-        public FigureTypes Bonus { get; } = FigureTypes.Nothing;
-        public FigureTypes AntiBonus { get; } = FigureTypes.Nothing;
-        public double FullHp { get; } = 100;
-        public double Attack { get; } = 100;
-        public double Defence { get; } = 0;
-        public bool MovingAttack { get; } = true;
-        public int Cost { get; } = 3;
+        {0, new Uri("pack://application:,,,/BattleChess3.SilmarillionFigures;component/Images/IrmoUngoliant1.png", UriKind.Absolute)},
+        {1, new Uri("pack://application:,,,/BattleChess3.SilmarillionFigures;component/Images/IrmoUngoliant2.png", UriKind.Absolute)},
+    };
 
-        public Dictionary<int, Uri> ImageUris { get; } = new Dictionary<int, Uri>
-        {
-            {0, new Uri("pack://application:,,,/BattleChess3.SilmarillionFigures;component/Images/IrmoUngoliant1.png", UriKind.Absolute)},
-            {1, new Uri("pack://application:,,,/BattleChess3.SilmarillionFigures;component/Images/IrmoUngoliant2.png", UriKind.Absolute)},
-        };
+    public void AttackAction(ITile from, ITile to, ITile[] board)
+        => to.KillFigure(board);
 
-        public void AttackAction(ITile from, ITile to, ITile[] board)
-            => to.KillFigure(board);
+    public void MoveAction(ITile from, ITile to, ITile[] board)
+        => from.MoveToPosition(to.Position, board);
 
-        public void MoveAction(ITile from, ITile to, ITile[] board)
-            => from.MoveToPosition(to.Position, board);
-
-        private readonly Position[][] _moveChain = 
-        {
-            new Position[] {(-2, 1)},
-            new Position[] {(-2, -1)},
-            new Position[] {(2, 1)},
-            new Position[] {(2, -1)},
-            new Position[] {(-1, -2)},
-            new Position[] {(1, -2)},
-            new Position[] {(-1, 2)},
-            new Position[] {(1, 2)},
-        };
-        public Position[][] GetMoveChains(Position position) => _moveChain;
-        
-        
-        private readonly Position[][] _attackChain = 
-        {
-            new Position[] {(-2, 1)},
-            new Position[] {(-2, -1)},
-            new Position[] {(2, 1)},
-            new Position[] {(2, -1)},
-            new Position[] {(-1, -2)},
-            new Position[] {(1, -2)},
-            new Position[] {(-1, 2)},
-            new Position[] {(1, 2)},
-        };
-        public Position[][] GetAttackChains(Position position) => _attackChain;
-    }
+    private readonly Position[][] _moveChain = 
+    {
+        new Position[] {(-2, 1)},
+        new Position[] {(-2, -1)},
+        new Position[] {(2, 1)},
+        new Position[] {(2, -1)},
+        new Position[] {(-1, -2)},
+        new Position[] {(1, -2)},
+        new Position[] {(-1, 2)},
+        new Position[] {(1, 2)},
+    };
+    public Position[][] GetMoveChains(Position position) => _moveChain;
+    
+    
+    private readonly Position[][] _attackChain = 
+    {
+        new Position[] {(-2, 1)},
+        new Position[] {(-2, -1)},
+        new Position[] {(2, 1)},
+        new Position[] {(2, -1)},
+        new Position[] {(-1, -2)},
+        new Position[] {(1, -2)},
+        new Position[] {(-1, 2)},
+        new Position[] {(1, 2)},
+    };
+    public Position[][] GetAttackChains(Position position) => _attackChain;
 }
