@@ -1,22 +1,21 @@
 ﻿using BattleChess3.Core.Model.Figures;
 
-namespace BattleChess3.SilmarillionFigures
-{
-    public class SilmarillionFigureGroup : IFigureGroup
-    {
-        public string ShownName => "Silmarillion";
+namespace BattleChess3.SilmarillionFigures;
 
-        public IFigureType[] FigureTypes => new IFigureType[]
-        {
-            ManweMelkor.Instance,
-            UlmoAncalagon.Instance,
-            AuleGothmog.Instance,
-            ElfOrc.Instance,
-            IrmoUngoliant.Instance,
-            NiennaBalrog.Instance,
-            OromeCarcharoth.Instance,
-            VardaSauron.Instance,
-            YavannaGlaurung.Instance,
-        };
-    }
+public class SilmarillionFigureGroup : IFigureGroup
+{
+    public string ShownName => "Silmarillion";
+
+    public IFigureType[] FigureTypes => new IFigureType[]
+    {
+        ManweMelkor.Instance,
+        UlmoAncalagon.Instance,
+        AuleGothmog.Instance,
+        ElfOrc.Instance,
+        IrmoUngoliant.Instance,
+        NiennaBalrog.Instance,
+        OromeCarcharoth.Instance,
+        VardaSauron.Instance,
+        YavannaGlaurung.Instance,
+    };
 }
